Fix invalid VOLUME syntax in upstream container
According to Docker's docs, VOLUME syntax can either be a JSON array,
with the elements quoted, or a bare list of space separated items. The
upstream for the sqitch container is using some kind of in-between
syntax (`VOLUME [/src]`) which seems to cause problems on some versions
of Docker.

Since the upstream doesn't actually do a whole lot (basically just
installs sqitch w/ cpan), I decided to inline it for expediency. This
also let me use `cpanm` to install sqitch which saves a lot of time when
rebuilding.

Signed-off-by: Daniel DeLeo <dan@chef.io>